refactor: move fetch logic to dashboard
